+++ src/services/tracking.rs
@@ -1,27 +1,27 @@
 use std::io::{BufRead, Read, Write};
-use std::sync::atomic::{AtomicUsize, Ordering};
+use std::sync::atomic::{AtomicU64, AtomicUsize, Ordering};
 use std::sync::Arc;
 
 #[derive(Clone)]
 pub struct BytesCounter {
-    bytes_amount: Arc<AtomicUsize>,
+    bytes_amount: Arc<AtomicU64>,
 }
 
 impl BytesCounter {
-    pub fn new(bytes: &Arc<AtomicUsize>) -> BytesCounter {
+    pub fn new(bytes: &Arc<AtomicU64>) -> BytesCounter {
         BytesCounter {
             bytes_amount: Arc::clone(bytes),
         }
     }
 
-    pub fn value(&self) -> usize {
+    pub fn value(&self) -> u64 {
         self.bytes_amount.load(Ordering::Relaxed)
     }
 }
 
 pub struct BytesCountingReader<R: Read> {
     inner_reader: R,
-    byte_count: Arc<AtomicUsize>,
+    byte_count: Arc<AtomicU64>,
 }
 
 impl<R: Read> BytesCountingReader<R> {
@@ -29,7 +29,7 @@ impl<R: Read> BytesCountingReader<R> {
     pub fn new(inner: R) -> Self {
         BytesCountingReader {
             inner_reader: inner,
-            byte_count: Arc::new(AtomicUsize::new(0)),
+            byte_count: Arc::new(AtomicU64::new(0)),
         }
     }
 
@@ -46,7 +46,8 @@ impl<R: Read> BytesCountingReader<R> {
 impl<R: Read> Read for BytesCountingReader<R> {
     fn read(&mut self, buf: &mut [u8]) -> std::io::Result<usize> {
         let bytes_read = self.inner_reader.read(buf)?;
-        self.byte_count.fetch_add(bytes_read, Ordering::Relaxed);
+        self.byte_count
+            .fetch_add(bytes_read as u64, Ordering::Relaxed);
         Ok(bytes_read)
     }
 }
@@ -57,7 +58,7 @@ impl<R: Read + BufRead> BufRead for BytesCountingReader<R> {
     }
 
     fn consume(&mut self, amt: usize) {
-        self.byte_count.fetch_add(amt, Ordering::Relaxed);
+        self.byte_count.fetch_add(amt as u64, Ordering::Relaxed);
         self.inner_reader.consume(amt);
     }
 }
@@ -65,7 +66,7 @@ impl<R: Read + BufRead> BufRead for BytesCountingReader<R> {
 /// A wrapper around a `Write` type that tracks the number of bytes written.
 pub struct BytesCountingWriter<W: Write> {
     inner: W,
-    byte_count: Arc<AtomicUsize>,
+    byte_count: Arc<AtomicU64>,
 }
 
 impl<W: Write> BytesCountingWriter<W> {
@@ -73,7 +74,7 @@ impl<W: Write> BytesCountingWriter<W> {
     pub fn new(inner: W) -> Self {
         BytesCountingWriter {
             inner,
-            byte_count: Arc::new(AtomicUsize::new(0)),
+            byte_count: Arc::new(AtomicU64::new(0)),
         }
     }
 
@@ -90,7 +91,8 @@ impl<W: Write> BytesCountingWriter<W> {
 impl<W: Write> Write for BytesCountingWriter<W> {
     fn write(&mut self, buf: &[u8]) -> std::io::Result<usize> {
         let bytes_written = self.inner.write(buf)?;
-        self.byte_count.fetch_add(bytes_written, Ordering::Relaxed);
+        self.byte_count
+            .fetch_add(bytes_written as u64, Ordering::Relaxed);
         Ok(bytes_written)
     }
